djangoapp/requirements.txt
django>=4.2.3,<4.3
psycopg2-binary>=2.9.6,<2.10
django-cors-headers>=4.2.0,<4.3 
djangorestframework>=3.14.0,<3.15
requests
pandas
openpyxl>=3.1.2,<3.2
xlrd==2.0.1